LayoutTests/platform/iphone-simulator/accessibility/file-upload-button.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../../resources/js-test-pre.js"></script>
<script>
var successfullyParsed = false;
</script>
</head>
<body id="body">

<input id="file-input" type="file"/>

<p id="description"></p>
<div id="console"></div>

<script>

    description("This tests that the label and value are correct for input file types.");

    if (window.accessibilityController) {

        var button1 = accessibilityController.accessibleElementById("file-input");
        shouldBe("button1.iphoneLabel", "'AXLabel: Choose File'");
        shouldBe("button1.iphoneValue", "'AXValue: no file selected'");
    }

    successfullyParsed = true;
</script>

<script src="../../../resources/js-test-post.js"></script>

</body>
</html>
