safari-600.7-branch/LayoutTests/platform/ios-sim/accessibility/file-upload-button-expected.txt


This tests that the label and value are correct for input file types.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS button1.iphoneLabel is 'AXLabel: Choose File'
PASS button1.iphoneValue is 'AXValue: no file selected'
PASS successfullyParsed is true

TEST COMPLETE
